LPS-155670 Remove the only getLibDir method usage
LIFERAY_SHIELDED_CONTAINER_LIB_PORTAL_DIR is set in JSPEngineShieldedContainerInitializer which uses ServletContext's built-in method getCanonicalPath to get the shielded container lib path. The prop is only in use when server start up. There is no need to calculate the dir path in PropsUtil. Only need to normalize it then use it to set LIFERAY_WEB_PORTAL_DIR